unit tests pass after refactoring.
