+++ ela/_version.py
@@ -5,4 +5,4 @@ Version.
 
 Doing it this way provides for access in setup.py and via __version__
 """
-__version__ = "0.6.2"
+__version__ = "0.6.3"
