+++ ela/spatial.py
@@ -2,7 +2,11 @@ import rasterio
 import numpy as np
 import pandas as pd
 
-from ela.textproc import EASTING_COL, NORTHING_COL, DEPTH_FROM_AHD_COL, DEPTH_FROM_COL, DEPTH_TO_AHD_COL, DEPTH_TO_COL
+from ela.textproc import EASTING_COL, NORTHING_COL, DEPTH_FROM_AHD_COL, DEPTH_FROM_COL, DEPTH_TO_AHD_COL, DEPTH_TO_COL, PRIMARY_LITHO_COL, PRIMARY_LITHO_NUM_COL, SECONDARY_LITHO_COL, GEOMETRY_COL
+
+KNN_WEIGHTING = 'distance'
+
+from sklearn import neighbors
 
 
 def read_raster_value(dem,band_1,easting,northing):
@@ -24,25 +28,26 @@ def read_raster_value(dem,band_1,easting,northing):
     else:
         return band_1[row,col]
 
-def raster_drill(row,dem, raster_grid):
-    easting=row[EASTING_COL]
-    northing=row[NORTHING_COL]
-    return read_raster_value(dem,raster_grid,easting,northing)
+# DEPRECATED should not be superseded by HeightDatumConverter
+# def raster_drill(row,dem, raster_grid):
+#     easting=row[EASTING_COL]
+#     northing=row[NORTHING_COL]
+#     return read_raster_value(dem,raster_grid,easting,northing)
 
-def raster_drill_df(df, dem, raster_grid):
-    return df.apply (lambda row: raster_drill(row,dem, raster_grid),axis=1)
+# def raster_drill_df(df, dem, raster_grid):
+#     return df.apply (lambda row: raster_drill(row,dem, raster_grid),axis=1)
 
-def add_ahd(lithology_df, data_raster, drop_na=False):
-    df = lithology_df.copy(deep=True)
-    data_grid = data_raster.read(1)
-    nd = data_raster.nodata
-    ahd = raster_drill_df(df, data_raster, data_grid)
-    ahd[ahd==nd] = np.nan
-    df[DEPTH_FROM_AHD_COL]=ahd-df[DEPTH_FROM_COL]
-    df[DEPTH_TO_AHD_COL]=ahd-df[DEPTH_TO_COL]
-    if drop_na:
-        df = df[pd.notna(df[DEPTH_TO_AHD_COL])]
-    return df
+# def add_ahd(lithology_df, data_raster, drop_na=False):
+#     df = lithology_df.copy(deep=True)
+#     data_grid = data_raster.read(1)
+#     nd = data_raster.nodata
+#     ahd = raster_drill_df(df, data_raster, data_grid)
+#     ahd[ahd==nd] = np.nan
+#     df[DEPTH_FROM_AHD_COL]=ahd-df[DEPTH_FROM_COL]
+#     df[DEPTH_TO_AHD_COL]=ahd-df[DEPTH_TO_COL]
+#     if drop_na:
+#         df = df[pd.notna(df[DEPTH_TO_AHD_COL])]
+#     return df
 
 def get_coords_from_gpd_shape(shp, colname='geometry'):
     p = shp[[colname]]
@@ -64,11 +69,16 @@ class HeightDatumConverter:
         crs (str, dict, or CRS): The coordinate reference system.
         transform (Affine instance): Affine transformation mapping the pixel space to geographic space.
     """
-    def __init__(self, dem_raster):
+    def __init__(self, dem_raster, easting_col=EASTING_COL, northing_col=NORTHING_COL, depth_from_ahd_col=DEPTH_FROM_AHD_COL, depth_to_ahd_col=DEPTH_TO_AHD_COL):
         """Initialize this with a coordinate reference system object and an affine transform. See rasterio.
         
-        Args:
+            Args:
+                easting_col (str): name of the data frame column for easting
+                northing_col (str): name of the data frame column for northing
+                depth_from_ahd_col (str): name of the data frame column for the height of the top of the soil column (ahd stands for for australian height datum, but not restricted)
+                depth_to_ahd_col (str): name of the data frame column for the height of the bottom of the soil column (ahd stands for for australian height datum, but not restricted)        
         """
+        self.dfcn = GeospatialDataFrameColumnNames(easting_col, northing_col, depth_from_ahd_col, depth_to_ahd_col)
         self.dem_raster = dem_raster
         self.data_grid = self.dem_raster.read(1)
 
@@ -83,6 +93,9 @@ class HeightDatumConverter:
         for i in range(len(x)):
             v[i] = read_raster_value(self.dem_raster, self.data_grid, x[i], y[i])
         return v
+
+    def raster_drill_df(self, df):
+        return self._raster_drill_df(df, self.dfcn.easting_col, self.dfcn.northing_col)
 
     def add_height(self, lithology_df, 
         depth_from_col=DEPTH_FROM_COL, depth_to_col=DEPTH_TO_COL, 
@@ -414,4 +427,197 @@ def surface_array(raster, x_min, y_min, x_max, y_max, grid_res):
 
 # mlab.outline()
 # mlab.show()
+
+
+
+class GeospatialDataFrameColumnNames(object):
+    """Operations on data frames with 3D spatial information of lithology logs. 
+    The purpose of this class is to adapt 'pyela' operations 
+    to different data without requiring renaming columns.
+
+    Attributes:
+        easting_col (str): name of the data frame column for easting
+        northing_col (str): name of the data frame column for northing
+        depth_from_ahd_col (str): name of the data frame column for the height of the top of the soil column (ahd stands for for australian height datum, but not restricted)
+        depth_to_ahd_col (str): name of the data frame column for the height of the bottom of the soil column (ahd stands for for australian height datum, but not restricted)
+    """
+
+    def __init__(self, easting_col=EASTING_COL, northing_col=NORTHING_COL, depth_from_ahd_col=DEPTH_FROM_AHD_COL, depth_to_ahd_col=DEPTH_TO_AHD_COL):
+        """Constructor, operations on data frames with 3D spatial information of lithology logs
+
+            Args:
+                easting_col (str): name of the data frame column for easting
+                northing_col (str): name of the data frame column for northing
+                depth_from_ahd_col (str): name of the data frame column for the height of the top of the soil column (ahd stands for for australian height datum, but not restricted)
+                depth_to_ahd_col (str): name of the data frame column for the height of the bottom of the soil column (ahd stands for for australian height datum, but not restricted)        
+        """
+        self.easting_col = easting_col
+        self.northing_col = northing_col
+        self.depth_from_ahd_col = depth_from_ahd_col
+        self.depth_to_ahd_col = depth_to_ahd_col
+
+    def lithologydata_slice_depth(self, df, slice_depth):
+        """
+        Subset data frame with entries at a specified AHD coordinate    
+
+            Args:
+                df (pandas data frame): bore lithology data  
+                slice_depth (float): AHD coordinate at which to slice the data frame for lithology observations 
+        
+            Returns:
+                a (view of a) data frame, a subset of the input data frame, 
+                entries intersecting with the specified slice depth
+        """
+        df_slice=df.loc[(df[self.depth_from_ahd_col] >= slice_depth) & (df[self.depth_to_ahd_col] <= slice_depth)]
+        return df_slice
+
+    # The following was spurred by trying to get more data in KNN cross-validation, but this may be a dubious method to increase the data pool. Park.
+    # def get_lithology_observations_between(df, bottom_ahd, top_ahd, column_name ):
+    #     """
+    #     Subset data frame with entries at a specified AHD coordinate, and with valid lithology information.
+
+    #         Args:
+    #             df (pandas data frame): bore lithology data  
+    #             bottom_ahd (float): bottom AHD coordinate of the slice to subset
+    #             top_ahd (float): top AHD coordinate of the slice 
+    #             column_name (str): name of the column with string information to use to strip entries with missing lithology information
+        
+    #         Returns:
+    #             a (view of a) data frame; a subset of the input data frame, 
+    #             entries intersecting with the specified slice depth
+    #     """
+    #     depth_from_colname=DEPTH_FROM_AHD_COL
+    #     depth_to_colname=DEPTH_TO_AHD_COL
+    #     df_slice=df.loc[(df[depth_from_colname] >= top_ahd) & (df[depth_to_colname] <= slice_depth)] # CAREFUL HERE about order and criteria... trickier than 2D slicing.
+    #     df_1=df_slice[np.isnan(df_slice[column_name]) == False]
+    #     return df_1
+
+    def get_lithology_observations_for_depth(self, df, slice_depth, column_name ):
+        """
+        Subset data frame with entries at a specified AHD coordinate, and with valid lithology information.
+
+            Args:
+                df (pandas data frame): bore lithology data  
+                slice_depth (float): AHD coordinate at which to slice the data frame for lithology observations 
+                column_name (str): name of the column with string information to use to strip entries with missing lithology information
+        
+            Returns:
+                a (view of a) data frame; a subset of the input data frame, 
+                entries intersecting with the specified slice depth
+        """
+        df_slice = self.lithologydata_slice_depth(df, slice_depth)
+        df_1 = df_slice[np.isnan(df_slice[column_name]) == False]
+        return df_1
+
+    def extract_bore_class_num(self, bore_log_df, column_name):
+        """Gets the columns easting, northing, primary lithology class number, AHD depth 'from' and 'to' from a bore data log
+
+            Args:
+                bore_log_df (pandas data frame): bore lithology data  
+                column_name (str): name of the column of interest e.g. lithology descriptions or classes        
+        """
+        xx = bore_log_df[self.easting_col].values
+        yy = bore_log_df[self.northing_col].values
+        ss = bore_log_df[column_name].values
+        zz_from = bore_log_df[self.depth_from_ahd_col].values
+        zz_to = bore_log_df[self.depth_to_ahd_col].values
+        return xx, yy, zz_from, zz_to, ss
+
+    def make_training_set(self, observations, column_name):
+        """Create a training set suitable for machine learning by e.g. scikit-learn out of a georeferenced data frame.
+
+            Args:
+                observations (pandas data frame): bore lithology data  
+                column_name (str): name of the column of interest e.g. lithology descriptions or classes        
+        """
+        # X = observations.as_matrix(columns=[EASTING_COL, NORTHING_COL])
+        X = observations[[self.easting_col, self.northing_col]].values
+        y = np.array(observations[column_name])
+        #NOTE: should I also do e.g.:
+        #shuffle_index = np.random.permutation(len(y))
+        #X, y = X[shuffle_index], y[shuffle_index]   
+        return (X, y)
+
+    def get_knn_model(self, df, column_name, slice_depth, n_neighbours):
+        """Train a K-nearest neighbours model for a given plane 
+
+        Args:
+            df (data frame): 
+            column_name (str): 
+            slice_depth (numeric): 
+            n_neighbours (int): 
+
+        Returns:
+            KNeighborsClassifier: trained classifier.
+        """
+        df_1 = self.get_lithology_observations_for_depth(df, slice_depth, column_name)
+        X, y = self.make_training_set(df_1, column_name)
+        if n_neighbours > len(df_1):
+            return None
+        else:
+            knn = neighbors.KNeighborsClassifier(n_neighbours, weights = KNN_WEIGHTING).fit(X, y)
+            return knn
+            
+    def class_probability_estimates_depth(self, df, column_name, slice_depth, n_neighbours, mesh_grid, func_training_set=None):
+        """Subset data frame with entries at a specified AHD coordinate
+
+            Args:
+                df (pandas data frame): bore lithology data  
+                slice_depth (float): AHD coordinate at which to slice the data frame for lithology observations
+                n_neighbours (int): number of nearest neighbours 
+                mesh_grid (tuple): coordinate matrices to interpolate over (numpy.meshgrid)
+                func_training_set (callable):  a function to processing the training set (e.g. completing dummy with dummy classes, other not present in the trainining set)
+
+            Returns:
+                a list of numpy arrays, shaped like the meshgrid.
+        """
+        df_1 = self.get_lithology_observations_for_depth(df, slice_depth, column_name)
+        X, y = self.make_training_set(df_1, column_name)
+        if not (func_training_set is None):
+            X, y = func_training_set(X, y)
+        knn = neighbors.KNeighborsClassifier(n_neighbours, weights = KNN_WEIGHTING).fit(X, y)
+        xx, yy = mesh_grid
+        class_prob = knn.predict_proba(np.c_[xx.ravel(), yy.ravel()])
+        n_classes = class_prob.shape[1]
+        probs = []
+        for i in range(n_classes):
+            p = class_prob[:,i].reshape(xx.shape)
+            probs.append(p)
+        return probs
+
+    def class_probability_estimates_depth_bbox(self, df, column_name, slice_depth, n_neighbours, geo_pd, grid_res = 100, func_training_set=None):
+        mesh_grid = create_meshgrid(geo_pd, grid_res)
+        return self.class_probability_estimates_depth(df, column_name, slice_depth, n_neighbours, mesh_grid, func_training_set)
+
+    def interpolate_lithologydata_slice_depth(self, df, column_name, slice_depth, n_neighbours, mesh_grid):
+        """Interpolate lithology data
+
+            Args:
+                df (pandas data frame): bore lithology data  
+                slice_depth (float): AHD coordinate at which to slice the data frame for lithology observations
+                n_neighbours (int): number of nearest neighbours 
+                mesh_grid (tuple): coordinate matrices to interpolate over (numpy.meshgrid)
+
+            Returns:
+                numpy array, predicted values over the grid.
+        """
+        knn = self.get_knn_model(df, column_name, slice_depth, n_neighbours)
+        return interpolate_over_meshgrid(knn, mesh_grid)
+
+    def interpolate_lithologydata_slice_depth_bbox(self, df, column_name, slice_depth, n_neighbours, geo_pd, grid_res = 100):
+        """Interpolate lithology data
+
+            Args:
+                df (pandas data frame): bore lithology data  
+                slice_depth (float): AHD coordinate at which to slice the data frame for lithology observations
+                n_neighbours (int): number of nearest neighbours 
+                geo_pd (geopandas df): vector of spatial data from which to get the bounds of interest (bounding box)
+                grid_res (int): grid resolution in m for x and y.
+
+            Returns:
+                numpy array, predicted values over the grid.
+        """
+        mesh_grid = create_meshgrid(geo_pd, grid_res)
+        return self.interpolate_lithologydata_slice_depth(df, column_name, slice_depth, n_neighbours, mesh_grid)
+
